requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
requests==2.31.0
python-dotenv==1.0.0
twilio==8.5.0
openai==1.3.0
google-generativeai==0.3.0
APScheduler==3.10.4